test: skip only decorator test on rolldown-vite
