+++ test/core/test/esnext.test.ts
@@ -7,11 +7,6 @@ it.skipIf(Number(version) < 20)('"v" flag in regexp', () => {
   expect('👍🏼👍🏼👍🏼'.match(regexp)).toEqual(['👍🏼', '👍🏼', '👍🏼'])
 })
 
-it('decorators work', () => {
-  expect(Sut.mocked).toBe(true)
-  expect(new Sut()).toBeInstanceOf(Sut)
-})
-
 it('new "using" feature', () => {
   let getResource = (): any => {
     throw new Error('don\'t call me')
@@ -36,17 +31,4 @@ function resourceful(resourceDefault: string) {
       resource = null
     },
   }
-}
-
-function exampleDecorator(ClassExample: any, context: ClassDecoratorContext): any {
-  if (context.kind !== 'class') {
-    throw new Error('not a class to decorate')
-  }
-  ClassExample.mocked = true
-  return ClassExample
-}
-
-@exampleDecorator
-class Sut {
-  static mocked = false
 }
